Clean up WindowPartition class
The WindowPartition class got too complicated with the introduction
of RANGE and GROUPS framing. This splits out the behavior of each
framing scheme into separate classes.
